backend/src/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.31
feedparser>=6.0.10
transformers>=4.35.0
torch>=2.0.0
anthropic>=0.39.0
google-genai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
textblob>=0.17.1
scikit-learn>=1.3.0
python-dotenv>=1.0.0
cachetools>=5.3.0
